apps/api/pyproject.toml
[project]
name = "sports-system-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.0",
    "sqlmodel>=0.0.22",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.0",
    "pyjwt[crypto]>=2.8.0",
    "apscheduler>=3.10.0,<4.0",
    "httpx>=0.28.0",
    "fastapi-users[sqlalchemy]>=15.0.5",
    "sse-starlette>=3.3.4",
    "slowapi>=0.1.9",
    "orjson>=3.11.8",
    "tenacity>=9.1.4",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.28.0",
]

